Move PrefixHandler and ScopeMatcher out of SPI package
