+++ modules/apps/oauth2-provider/oauth2-provider-scopes-impl/src/main/java/com/liferay/oauth2/provider/scopes/impl/prefixhandler/BundleNamespacePrefixHandlerFactory.java
@@ -14,7 +14,7 @@
 
 package com.liferay.oauth2.provider.scopes.impl.prefixhandler;
 
-import com.liferay.oauth2.provider.scopes.spi.PrefixHandler;
+import com.liferay.oauth2.provider.scopes.prefixhandler.PrefixHandler;
 import com.liferay.oauth2.provider.scopes.spi.PrefixHandlerFactory;
 import com.liferay.oauth2.provider.scopes.spi.PropertyGetter;
 import com.liferay.portal.kernel.util.MapUtil;
