+++ modules/apps/oauth2-provider/oauth2-provider-scopes-impl/src/test/java/com/liferay/oauth2/provider/scopes/impl/ChunkScopeMatcherFactoryTest.java
@@ -16,9 +16,9 @@ package com.liferay.oauth2.provider.scopes.impl;
 
 import com.liferay.oauth2.provider.scopes.impl.prefixhandler.DefaultPrefixHandlerFactory;
 import com.liferay.oauth2.provider.scopes.impl.scopematcher.ChunkScopeMatcherFactory;
-import com.liferay.oauth2.provider.scopes.spi.ScopeMatcher;
+import com.liferay.oauth2.provider.scopes.prefixhandler.PrefixHandler;
+import com.liferay.oauth2.provider.scopes.scopematcher.ScopeMatcher;
 import com.liferay.oauth2.provider.scopes.spi.ScopeMatcherFactory;
-import com.liferay.oauth2.provider.scopes.spi.PrefixHandler;
 import com.liferay.portal.kernel.util.StringPool;
 import org.junit.Test;
