+++ modules/apps/oauth2-provider/oauth2-provider-scopes-spi/bnd.bnd
@@ -2,4 +2,4 @@ Bundle-Name: Liferay OAuth2 Provider Scopes SPI
 Bundle-SymbolicName: com.liferay.oauth2.provider.liferay.spi
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.oauth2.provider.scopes.spi
+	com.liferay.oauth2.provider.scopes.*
